Issue #11163: Enforce file size for LocalFinalVariableName input files
