+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/naming/localfinalvariablename/InputLocalFinalVariableNameEnum2.java
@@ -1,0 +1,22 @@
+/*
+LocalFinalVariableName
+format = [A-Z]+
+tokens = (default)VARIABLE_DEF, PARAMETER_DEF, RESOURCE
+
+
+*/
+
+package com.puppycrawl.tools.checkstyle.checks.naming.localfinalvariablename;
+
+/** Test enum for member naming check */
+enum InputLocalFinalVariableNameEnum2
+{
+    /** ABC constant */
+    ABC,
+
+    /** XYZ constant */
+    XYZ;
+
+    /** Should be mSomeMember */
+    private int someMember;
+}
